requirements.txt
langchain-core>=0.1.0
google-genai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.2
aiohttp>=3.8.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
pymongo>=4.6.0
yfinance>=0.2.40
pandas>=2.0.0
langchain-perplexity >= 0.1.1
